Added namecoin domains
Not sure if multiple DNS servers are allowed like that, but since it's a list I assume they are. That's all the active ones listed on [this page](http://dot-bit.org/HowToBrowseBitDomains#List_of_DNS_servers) at this time. Expect more alternate DNS-roots in the future (as long as they don't conflict)